+++ src/activitypub/actors/mod.rs
@@ -1,3 +1,4 @@
 mod attachments;
 pub mod helpers;
+mod keys;
 pub mod types;