+++ src/activitypub/constants.rs
@@ -6,6 +6,7 @@ pub const AS_MEDIA_TYPE: &str = "application/activity+json";
 pub const AP_CONTEXT: &str = "https://www.w3.org/ns/activitystreams";
 pub const W3ID_SECURITY_CONTEXT: &str = "https://w3id.org/security/v1";
 pub const W3ID_DATA_INTEGRITY_CONTEXT: &str = "https://w3id.org/security/data-integrity/v1";
+pub const W3ID_MULTIKEY_CONTEXT: &str = "https://w3id.org/security/multikey/v1";
 pub const SCHEMA_ORG_CONTEXT: &str = "http://schema.org/";
 pub const MASTODON_CONTEXT: &str = "http://joinmastodon.org/ns#";
 pub const MITRA_CONTEXT: &str = "http://jsonld.mitra.social#";